requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-jose[cryptography]==3.3.0
httpx==0.27.0
redis[async]==5.0.7
pytest==8.2.2
pytest-asyncio==0.23.8
anyio==4.4.0
